requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
pandas
streamlit
langchain_huggingface
requests>=2.28.0
python-dotenv>=0.21.0